Fix vault migration layer boundary
